Allowing arbitrary parameters to be passed to the API payload.
Push notifications, both iOS and Android, can have arbitrary parameters associated with them. To pass such an arbitrary parameter, use a key that isn't a part of the Urban Airship API, and it will be shipped with the push notification to the devices.

This is useful for passing arbitrary metadata along with the notification itself, such as the numeric ID of the record associated with a push notification, if any.

The documentation for this feature in the Urban Airship API can be found here: https://support.urbanairship.com/customer/portal/articles/60532-sending-custom-push-data