Only use Ruby extensions if ActiveSupport missing
Overriding ActiveSupport's `Object#try` and `Hash#Compact` was
surprising behavior to find inside a gem. It's possible that when in a
Rails environment, the load order will use this gem's implementation of
the two methods rather than the well tested solutions in ActiveSupport.

This commit proposes only defining these monkeypatches in the event that
the methods have _not_ already been defined.