Updates alias to coerce to string
The API appears to require that the "alias" value in the payload be a
string. Unfortunately if a string is passed, you will receive an error
similar to the following:

```
{
  "ok" => false,
  "error" => "Could not parse request body.",
  "error_code" => 40214,
  "details" => {
    "error"=>"Empty selector."
  },
  ...
}
```

This commit resolves that issue by coercing the parameter passed to the
`alias` method into a string before including it into the payload.